Introduce `LightweightJsonTestScriptTestRunner`
