+++ tests/phpunit/Utils/JSONScript/QueryTestCaseProcessor.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace SMW\Tests\Integration\JSONScript;
+namespace SMW\Tests\Utils\JSONScript;
 
 use SMW\Query\Parser as QueryParser;
 use SMW\Store;
