+++ tests/phpunit/Utils/JSONScript/RdfTestCaseProcessor.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace SMW\Tests\Integration\JSONScript;
+namespace SMW\Tests\Utils\JSONScript;
 
 use SMWExportController as ExportController;
 use SMWRDFXMLSerializer as RDFXMLSerializer;
